+++ src/java/Changes.md
@@ -4,6 +4,7 @@
 
 * BREAKING: The DXAPI class may no longer be instantiated. You must use its
   methods statically.
+* New classes: DXObject, DXProject, and DXContainer
 
 ## 0.74.0